add graphcast model and example
