+++ ppsci/arch/__init__.py
@@ -16,6 +16,7 @@ from __future__ import annotations
 
 import copy
 
+from ppsci.arch.graphcast import GraphCastNet
 from ppsci.arch.phycrnet import PhyCRNet
 
 from ppsci.arch.base import Arch  # isort:skip
@@ -70,6 +71,7 @@ __all__ = [
     "AutoEncoder",
     "build_model",
     "CFDGCN",
+    "GraphCastNet",
 ]
